+++ actions/get_build_number.py
@@ -1,7 +1,4 @@
-try:  # Python 3
-    from http import HTTPStatus as http_status
-except ImportError:  # Python 2
-    import httplib as http_status
+import six.moves.http_client as http_status
 
 from lib.action import CircleCI
 
@@ -19,8 +16,8 @@ class GetBuildNumberAction(CircleCI):
             extra_headers={'limit': str(search_limit)}
         )
 
-        if response.status_code != http_status.OK:
-            raise Exception('Project %s not found.' % project)
+        if response.status_code != http_status.OK:  # pylint: disable=no-member
+            raise Exception('Project %s not found. Response: %s' % (project, response.text))
 
         for build in response.json():
             if build['vcs_revision'] == vcs_revision: