+++ actions/retry_build.py
@@ -1,7 +1,4 @@
-try:  # Python 3
-    from http import HTTPStatus as http_status
-except ImportError:  # Python 2
-    import httplib as http_status
+import six.moves.http_client as http_status
 
 from lib.action import CircleCI
 
@@ -18,7 +15,7 @@ class RetryBuild(CircleCI):
             path, method='POST'
         )
 
-        if response.status_code != http_status.CREATED:
+        if response.status_code != http_status.CREATED:  # pylint: disable=no-member
             raise Exception('Project %s not found.' % project)
 
         return response.json()