+++ actions/wait_until_build_finishes.py
@@ -1,8 +1,6 @@
-try:  # Python 3
-    from http import HTTPStatus as http_status
-except ImportError:  # Python 2
-    import httplib as http_status
 import time
+
+import six.moves.http_client as http_status
 
 from lib.action import CircleCI
 
@@ -22,7 +20,7 @@ class WaitUntilBuildFinishes(CircleCI):
                 path, method='GET',
             )
 
-            if response.status_code != http_status.OK:
+            if response.status_code != http_status.OK:  # pylint: disable=no-member
                 msg = ('Build number %s for project ' % build_num +
                        '%s not found.' % project)
                 raise Exception(msg)